updates for two keyboards
planck:
added rgblight idle timeout.
renamed lower layer nether, and created a new lower layer; mainly
for shift + num row keys.

hs60:
just a rename of rgblight_on to rgblight_idle (and reverse it's meaning)
to match implementation for planck, which is a better representation of
its usage and intent.